+++ Practical-10/middleware/notAuth.js
@@ -1,6 +1,6 @@
 export default function ({ store, redirect }) {
   if (store.getters["auth/isAuthenticated"]) {
     console.log("Authenticated");
-    return redirect("/");
+    return redirect("/HomePage");
   }
 }
