+++ Practical-10/middleware/redirect.js
@@ -1,0 +1,6 @@
+export default async function (context) {
+  const redirectUrl = await getUrlForRedirect(context.route.path);
+  if (redirectUrl) {
+    context.redirect(301, redirectUrl);
+  }
+}
