+++ Practical-10/store/index.js
@@ -4,16 +4,6 @@ import auth from "./modules/auth/index.js";
 
 const createStore = () => {
   return new Vuex.Store({
-    // nuxtServerInit({ commit }) {
-    //   if (this.getters["auth/isAuthenticated"]) {
-    //     console.log("Authenticated.....");
-    //     this.$router.push("/");
-    //   } else {
-    //     console.log("Not Authenticated....");
-    //     this.$router.push("/userAuth");
-    //   }
-    // },
-
     modules: {
       cars,
       auth,
